Some small fixes to this unit test file to make it work properly if run without the central test suite file.
Still unable to get testRetrievePackageSources to run without an error, it can't
find the package 'fortune' as it does not apply the 'ruby -I./packages' search path.
This apperantly needs to be added by the initial ruby file which calls ruby with its 
interpreter options. From the unit testing framework this call is made before the 
TestAbtDownloadManager.rb file is called so the option has no effect here. No solution 
found yet...