Implemented log managers cache package.
Refactored abt install to make use of this automatically after a successful install.

Expanded the packages path to now be on the system. (I have just symlinked this location to my svn checkout root location)

Changed helper message when a package is called for installation, but is all ready installed. In this case it suggests trying reinstalling it.

Another test passing, 26 to go, all unit tests for AbTLogManager are now passing!